reorganizing gui files
